Pull boundary validation out into its own file
