+++ src/components/nodes/boundaryEscalationEvent/boundaryEscalationEvent.vue
@@ -1,7 +1,7 @@
 <script>
 import BoundaryEvent from '@/components/nodes/boundaryEvent/boundaryEvent';
 import escalationIcon from '@/assets/boundary-escalation-icon.svg';
-import { isValidBoundaryEscalationEvent } from '@/targetValidationUtils';
+import { isValidBoundaryEscalationEvent } from '@/boundaryEventValidation';
 
 export default {
   extends: BoundaryEvent,
